Basic test for Activity model. Installed sequelize-cli. TODO: clear database after tests, move config variables somewhere
