+++ package.json
@@ -28,7 +28,8 @@
     "pg": "^7.4.3",
     "pg-hstore": "^2.3.2",
     "request": "^2.88.0",
-    "sequelize": "^4.38.1"
+    "sequelize": "^4.38.1",
+    "sequelize-cli": "^4.1.1"
   },
   "devDependencies": {
     "eslint": "^5.5.0",
